+++ tests/testthat/test_get_all_typeforms.R
@@ -1,5 +1,7 @@
 test_that("Testing get_all_typeforms", {
   skip_on_cran()
+  if(nchar(Sys.getenv("typeform_api")) == 0) return(invisible(TRUE))
+
   typeforms = get_all_typeforms()
   expect_equal(ncol(typeforms$content), 2)
 }