+++ NEWS.md
@@ -1,3 +1,11 @@
+# Version 0.3.0
+  * Breaking changes to the API. `get_results()` now depreciated. 
+  Instead, use `get_questionnaire()`. This returns a list with http_status, 
+    question stats, questions, completed, and uncompleted responses (thanks to @hrbrmstr).
+
+# Version 0.2.1
+  * Adding automatic type conversion in the `get_results()` function (thanks to @1beb).
+
 # Version 0.2.0
   * Adding user agent to API call.
   * Better error handling.